pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Reusable infrastructure utilities (database connectors, Google Sheets, etc.)"
requires-python = ">=3.12"
dependencies = [
    "azure-core>=1.35.1",
    "azure-identity>=1.25.1",
    "mssql-python>=1.1.0",
    "mysql-connector",
    "sshtunnel>=0.4.0",
    "paramiko==3.5.1",
    "pygsheets>=2.0.6",
    "pandas==2.1.4",
    "numpy==1.26.4",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
